+++ src/components/pages/educators/index.js
@@ -1,0 +1,3 @@
+import SignUpModal from './sign-up-modal';
+
+export { SignUpModal };
